+++ lib/mdex/types/options.ex
@@ -9,7 +9,13 @@ defmodule MDEx.Types.ExtensionOptions do
             header_ids: nil,
             footnotes: false,
             description_lists: false,
-            front_matter_delimiter: nil
+            front_matter_delimiter: nil,
+            multiline_block_quotes: false,
+            math_dollars: false,
+            math_code: false,
+            shortcodes: false,
+            wikilinks_title_after_pipe: false,
+            wikilinks_title_before_pipe: false
 end
 
 defmodule MDEx.Types.ParseOptions do
@@ -29,7 +35,8 @@ defmodule MDEx.Types.RenderOptions do
             unsafe_: false,
             escape: false,
             list_style: :dash,
-            sourcepos: false
+            sourcepos: false,
+            escaped_char_spans: false
 end
 
 defmodule MDEx.Types.FeaturesOptions do